Update patch set 2
Topic set to "rotation test when HDMI is plugged".

This CTS test case set "Settings.System.USER_ROTATION" as ROTATION_90, and set "Settings.System.ACCELEROMETER_ROTATION" as 0. 

In this case, APPs which support both Landscape & Portriatm mode will rotate to ROTATION_90 without HDMI plugged. But when HDMI is plugged, CTS will find that it is still in Landscape mode then cause a failure. 

So if this CTS can check if HDMI is plugged and accept the different orientation result. It should be pass.

Patch-set: 2