Center child View of a ListView row within the ListView
Views that appear "below the fold" of a ListView row when that row is
taller than the listview itself cannot be clicked on and rightly show
as !isDisplayed(). This fix ensures that those views appear within
the ListView by centering the child View in question within the
ListView. A new TallRowAdapterDataIntegrationTest is included to show
the error. All other espresso tests work on my device with this patch
installed.

Change-Id: Ibca83e933fa2889adee5e0e14d1fcf9f245d02d0