Add failing test to ensure Intents is initialized
The IntentsTestRule only initializes Intents class after the Activity it is launching has successfully launched. If this Activity fails to launch (if for example it times out), then IntentsTestRule will still try to release Intents. This will result in NPE. This is a failing test that will confirm the problem and ensure when it is fixed.

Change-Id: Iaa40b7f117a56b9fbf0570087817e69cb105c226
Signed-off-by: Jason Atwood <jatwood@bignerdranch.com>