Update patch set 1
Patch Set 1:

Hi Stephan,

Thanks for quick response. I would like to mention that this action is not written for synchronization purpose which Espresso is very good at. This action is meant for the views that change state on performing same action repeatedly and your test wants to proceed further only on reaching a particular state in that view. For example, view pagers on which swiping changes the view state(e.g. swiping across images). Other examples can be calendar views where clicking the right/left navigation buttons changes the month and you are waiting for the expected month-year to appear on the screen. These kind of testing is needed where the view state is not fixed and changes from one run to another(again calendar is such an example).

I would also like to mention that we are already using such action in our project where we are showing animated promos to the user, where promos keep sliding automatically after regular intervals and there is an option to slide them by swiping manually as well, and we want to reach a particular promo and click on that to proceed further in our test. Now because promos keep sliding automatically so we want to perform swipe action repeatedly here to reach the desired promo view, otherwise our test was becoming flaky.

I hope this answers your question. Please let me know if you want me to elaborate more.

Patch-set: 1
Label: Verified=0