Enhancing MonitoringInstrumentation that enables the users to provide instance of Activity under test
Added a couple of methods in MonitoringInstrumentation that can be used to provide InterceptingActivityFactory and reset it. An InterceptingActivityFactory can be used to create an Activity object provided it supports creating it. If set, InterceptingActivityFactory is used by MonitoringInstrumentation to return Activity object on call of newActivity() method, otherwise default mechanism is used to create Activity object.
Also added a constructor in ActivityTestRule that accepts SingleActivityFactory, this SingleActivityFactory is set in the instrumentation before test is run and reset in the teardown. This provides a facility of testing the activity in isolation with the outer world by overriding the methods e.g. startService(), sendBroadcast() etc to do whatever test wants to do.

Change-Id: Ic67dfa4d1bee803854a1f6583998ccdb48e3cce2