Update patch set 1
Patch Set 1:

(7 comments)

Thanks again for another contribution! Looks like a great addition. I added some initial comments. Before we review this further, can you please add a few integration tests for this in the Espresso testapp which uses the Activity Provider in ATR and showcases a few scenarios like mocking a service, content resolver or scenarios this could be useful for? One last note, the current code for ATR in AOSP is a bit outdated, so we if we decide to merge this in we have to agree on what the best approach would be first. Thanks.

Patch-set: 1
Label: Presubmit-Verified=0