Update patch set 2
Patch Set 2: Verified+1 Code-Review+2

android.support.test.annotation was not pushed to open source yet. We will take care of it when we merge this in.

I just +2'd this CL and verified it, so no more work on your end. But we decided not to merge it in just yet and submit once we prepare for our next release.

Patch-set: 2
Reviewer: Stephan Linzner <1040223@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2
Label: Verified=+1