Added ViewAssertion to verify view's UI appearance
This will enable writing tests where UI can be tested not only for view's properties
but also its appearance for a particular form-factor. This feature is meant for tests
that run only on a particular device or form-factor. This feature can also be used to
test whole screen at once by using onView(isRoot()). Following are the tasks that were
done as part of this commit -

 - Added a method matchesVisually() in ViewAssertions
 - Extracted out Bitmap comparison logic in BitmapUtil class which is shared by HasBackgroundMatcher and matchesVisually()
 - Added IntegrationTest for verifying behavior of matchesVisually() on a sample view

Change-Id: I1cfa04a39f0e939ce1daa461bc45cb1280ad34e9