Update patch set 1
Patch Set 1: Code-Review-1

(5 comments)

Hi Prateek, thank for the contribution! sorry for the delay, I've actually looked at this change few weeks back and realized that we already have a design to implement a similar feature in a more generic way. 

I'm not really sure how to proceed with this change. Unfortunately, we're tied with other pressing issues. So we need some time to make a decision of how and who should implement this. When we get around to it I will post an update here. 

(While going through this change I've made a bunch of comments as I went along so I will just post them here but like I said we're going to put this change on pause for now)

Patch-set: 1
Label: Code-Review=-1