Update patch set 2
Patch Set 2: Code-Review-1

Sorry for the delay on this Prateek, I'm closing this change since it cannot go in as is. It requires a proper design document evaluating different approaches. I would recommend to create a feature request bug for this and move the discussion there. I would also recommend to link this cahnge to the bug for reference.

Patch-set: 2
Reviewer: Nick Korostelev <1052388@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1